pyproject.toml
[project]
name = "tp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "jesse tweedle", email = "jesse.tweedle@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "microtrade-fwf>=0.2.4",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
tp = "tp.pipeline:main"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
